modified return of ft_error_msg

+++ essencial_funcs/simple_funcs/ft_error_msg.c
@@ -6,13 +6,13 @@
 /*   By: ericlazo <erlazo@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/19 16:11:03 by ericlazo          #+#    #+#             */
-/*   Updated: 2020/09/19 16:16:01 by ericlazo         ###   ########.fr       */
+/*   Updated: 2020/09/23 15:08:26 by ericlazo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_error_msg(char *str, int ret)
+long	ft_error_msg(char *str, int ret)
 {
 	if (!str)
 		return (ret);
